eval based on length

+++ python/util/parameters.py
@@ -49,7 +49,7 @@ parser.add_argument("--test", action='store_true', help="Call if you want to onl
 
 args = parser.parse_args()
 
-
+"""
 # Check if test sets are available. If not, create an empty file.
 test_matched = "{}/multinli_0.9/multinli_0.9_test_matched_unlabeled.jsonl".format(args.datapath)
 
@@ -59,6 +59,20 @@ if os.path.isfile(test_matched):
     test_path = "{}/multinli_0.9/".format(args.datapath)
 else:
     test_path = "{}/multinli_0.9/".format(args.datapath)
+    temp_file = os.path.join(test_path, "temp.jsonl")
+    io.open(temp_file, "wb")
+    test_matched = temp_file
+    test_mismatched = temp_file
+"""
+# Check if test sets are available. If not, create an empty file.
+test_matched = "{}/multinli_0.9/multinli_0.9_test_matched.jsonl".format(args.datapath)
+
+if os.path.isfile(test_matched):
+    test_matched = "{}/multinli_0.9/multinli_0.9_dev_matched.jsonl".format(args.datapath) #"{}/multinli_0.9/multinli_0.9_test_matched.jsonl".format(args.datapath)
+    test_mismatched = "{}/multinli_0.9/multinli_0.9_dev_mismatched.jsonl".format(args.datapath) #"{}/multinli_0.9/multinli_0.9_test_mismatched.jsonl".format(args.datapath)
+    test_path = "{}".format(args.datapath)
+else:
+    test_path = "{}".format(args.datapath)
     temp_file = os.path.join(test_path, "temp.jsonl")
     io.open(temp_file, "wb")
     test_matched = temp_file
@@ -78,11 +92,14 @@ def load_parameters():
         "dev_snli": "{}/snli_1.0/snli_1.0_dev.jsonl".format(args.datapath),
         "test_snli": "{}/snli_1.0/snli_1.0_test.jsonl".format(args.datapath),
         "embedding_data_path": "{}/glove.840B.300d.txt".format(args.datapath),
+        #"embedding_data_path": "{}/glove.6B.50d.txt".format(args.datapath),
         "log_path": "{}".format(args.logpath),
         "ckpt_path":  "{}".format(args.ckptpath),
         "embeddings_to_load": args.emb_to_load,
         "word_embedding_dim": 300,
         "hidden_embedding_dim": 300,
+        #"word_embedding_dim": 50,
+        #"hidden_embedding_dim": 50,
         "seq_length": args.seq_length,
         "keep_rate": args.keep_rate, 
         "batch_size": 32,
